Settings Files and Form
Moved from the custom settings class to the Settings.Settings file. The settings form is now accessible from the launcher. Settings form loads the values stored in memory upon startup. Also implemented a button that puts the determined default values in the text boxers if the user wishes to reset.