Text Box Focus Problem Temp Fix
There was a problem where clicking on the textboxes in certain areas
would make the entire window lose focus. I still do not know what causes
it, but I found a temp fix by arranging the buttons, labels and textboxes in areas where you can click on them with no problems. The labels and buttons are now stacked vertically, and now have a complete PaleGreen bacground. Some areas of the labels still cause window to lose focus when clicked on.